configure-module: add `LOCALNETWORKS` env variable
The variable is used to determine if a device is present on the local
network. If the device is on the local network, Kamailio must present
itself to the device with its private IP address. If the device is
located outside on the internet, Kamailio must use its public IP
address.

Additional local networks can be added using the `local_networks` field
of the JSON object passed to the `configure-module` action.
The field is only used in a NAT scenario when a public IP is also set.